+++ src/components/line/filterToolbox/dimensions.js
@@ -1,32 +1,19 @@
-import React, { useMemo } from "react"
-import Menu from "@netdata/netdata-ui/lib/components/drops/menu"
+import React from "react"
 import { ItemContainer } from "@netdata/netdata-ui/lib/components/drops/menu/dropdownItem"
 import { TextSmall } from "@netdata/netdata-ui/lib/components/typography"
 import checkmark from "@netdata/netdata-ui/lib/components/icon/assets/checkmark_s.svg"
 import { Checkbox } from "@netdata/netdata-ui/lib/components/checkbox"
 import { useChart, useAttributeValue, useMetadata } from "@/components/provider"
 import Icon from "@/components/icon"
-import Label from "./label"
-
-const getItems = dimensions =>
-  dimensions
-    ? ["all", ...Object.keys(dimensions)].map(value => ({
-        label: dimensions[value]?.name || value,
-        value,
-      }))
-    : [{ value: "all" }]
-
-const getLabel = (value, items) => {
-  if (value.length === 0) return `All dimensions`
-  if (value.length === 1) {
-    const item = items.find(({ value: itemValue }) => value[0] === itemValue)
-    return item?.label || value[0]
-  }
-  return `${value.length} dimensions`
-}
+import Multiselect from "./multiselect"
 
 const CheckboxIcon = props => {
   return <Icon svg={checkmark} {...props} size="16px" />
+}
+
+const tooltipProps = {
+  heading: "Dimensions",
+  body: "Select one, multiple or all dimensions. A dimension is any value, either raw data or the result of a calculation that Netdata visualizes on a chart.",
 }
 
 const iconProps = { as: CheckboxIcon }
@@ -52,46 +39,22 @@ const renderItem = props => {
   return <Item key={key} {...props} />
 }
 
-const tooltipProps = {
-  heading: "Dimensions",
-  body:
-    "Select one, multiple or all dimensions. A dimension is any value, either raw data or the result of a calculation that Netdata visualizes on a chart.",
-}
-
 const Dimensions = ({ labelProps, ...rest }) => {
   const chart = useChart()
-
   const value = useAttributeValue("dimensions")
   const { dimensions } = useMetadata()
 
-  const options = useMemo(() => getItems(dimensions), [value, dimensions])
-
-  const label = getLabel(value, options)
-
   return (
-    <Menu
-      value={value}
-      onChange={chart.updateDimensionsAttribute}
-      items={options}
-      renderItem={renderItem}
-      closeOnClick={false}
+    <Multiselect
+      attrName="dimensions"
+      allName="All dimensions"
       data-track={chart.track("dimensions")}
-      dropProps={{
-        height: { max: "460px" },
-        overflow: "auto",
-        align: { top: "bottom", left: "left" },
-        "data-toolbox": true,
-      }}
+      options={dimensions}
+      renderItem={renderItem}
+      tooltipProps={tooltipProps}
+      value={value}
       {...rest}
-    >
-      <Label
-        secondaryLabel="select"
-        label={label}
-        title={tooltipProps.heading}
-        tooltipProps={tooltipProps}
-        {...labelProps}
-      />
-    </Menu>
+    />
   )
 }
